script: Move url interfaces into url/
