+++ components/script/dom/mod.rs
@@ -406,9 +406,7 @@ pub(crate) mod trustedtypes;
 pub(crate) use self::trustedtypes::*;
 pub(crate) mod uievent;
 pub(crate) mod url;
-pub(crate) mod urlhelper;
-pub(crate) mod urlpattern;
-pub(crate) mod urlsearchparams;
+pub(crate) use self::url::*;
 pub(crate) mod useractivation;
 pub(crate) mod userscripts;
 pub(crate) mod validation;
