+++ components/script/dom/testing/testbinding.rs
@@ -51,7 +51,7 @@ use crate::dom::globalscope::GlobalScope;
 use crate::dom::node::Node;
 use crate::dom::promise::Promise;
 use crate::dom::promisenativehandler::{Callback, PromiseNativeHandler};
-use crate::dom::url::URL;
+use crate::dom::url::url::URL;
 use crate::realms::InRealm;
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
 use crate::timers::OneshotTimerCallback;
